added earlier connection test
Added an "UNABLE TO CONNECT" message test when sending "0100" after initial auto-protocol (ATSP0). [line 221]

My ELM device (STN11xx) gave an immediate response of "SEARCHING...\rUNABLE TO CONNECT\r\r>", when it wasn't connected to an ECU.

elm327.py would then go through the ATTPx commands and fail on the first one with "CAN ERROR", but this was counted as a success since the response wasn't "UNABLE TO CONNECT". Ultimately, failure would occur when sending for supported commands.